sort the files into folders
