+++ examples/coffeemaker/src/test/java/edu/pjatk/inn/coffeemaker/CoffeeMakingTest.java
@@ -136,4 +136,3 @@ public class CoffeeMakingTest {
 	}
 
 }
-
